move `get_fingerprint_type` to fingerprinting utils module
